+++ src/renderer/hooks/dialog/useOpenViewGroupDialog.ts
@@ -1,5 +1,4 @@
 import { useCallback } from 'react'
-import { C } from '@deltachat/jsonrpc-client'
 
 import useDialog from './useDialog'
 import ViewGroup from '../../components/dialogs/ViewGroup'
@@ -13,7 +12,6 @@ export default function useOpenViewGroupDialog() {
     (chat: T.FullChat) => {
       openDialog(ViewGroup, {
         chat,
-        isBroadcast: chat.chatType === C.DC_CHAT_TYPE_BROADCAST,
       })
     },
     [openDialog]